cpp: bioformats: Move TileCache out from tiff namespace
This is also usable in other readers and writers.
