+++ cpp/lib/ome/bioformats/CMakeLists.txt
@@ -49,6 +49,7 @@ set(OME_BIOFORMATS_SOURCES
     PixelBuffer.cpp
     PixelProperties.cpp
     TileBuffer.cpp
+    TileCache.cpp
     UnknownFormatException.cpp
     UnsupportedCompressionException.cpp
     VariantPixelBuffer.cpp
@@ -71,6 +72,7 @@ set(OME_BIOFORMATS_HEADERS
     PixelBuffer.h
     PixelProperties.h
     TileBuffer.h
+    TileCache.h
     Types.h
     UnknownFormatException.h
     UnsupportedCompressionException.h
@@ -102,7 +104,6 @@ set(OME_BIOFORMATS_TIFF_SOURCES
     tiff/Sentry.cpp
     tiff/Tags.cpp
     tiff/TIFF.cpp
-    tiff/TileCache.cpp
     tiff/TileCoverage.cpp
     tiff/TileInfo.cpp)
 
@@ -114,7 +115,6 @@ set(OME_BIOFORMATS_TIFF_HEADERS
     tiff/Sentry.h
     tiff/Tags.h
     tiff/TIFF.h
-    tiff/TileCache.h
     tiff/TileCoverage.h
     tiff/TileInfo.h
     tiff/Types.h)
