+++ cpp/test/ome-bioformats/tilecache.cpp
@@ -38,13 +38,12 @@
 
 #include <ome/bioformats/Types.h>
 #include <ome/bioformats/TileBuffer.h>
-#include <ome/bioformats/tiff/Types.h>
-#include <ome/bioformats/tiff/TileCache.h>
+#include <ome/bioformats/TileCache.h>
 
 #include <ome/test/test.h>
 
 using ome::bioformats::dimension_size_type;
-using ome::bioformats::tiff::TileCache;
+using ome::bioformats::TileCache;
 using ome::bioformats::TileBuffer;
 
 TEST(TileCache, Construct)
